ocr_system/requirements.txt
pytesseract==0.3.10
Pillow==10.0.1
pdf2image==1.16.3
opencv-python==4.8.1.78
PyPDF2==3.0.1
pdfplumber==0.9.0
python-docx==0.8.11
numpy==1.24.3
